pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.84.0",
    "flask>=3.1.3",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "stripe>=14.4.0",
]
